start theme cleanup
