+++ src/ploneintranet/layout/viewlets/configure.zcml
@@ -19,4 +19,13 @@
         layer="ploneintranet.theme.interfaces.IThemeSpecific"
         />
 
+    <browser:viewlet
+        name="plone.personal_bar"
+        manager="plone.app.layout.viewlets.interfaces.IToolbar"
+        class=".personalbar.PersonalBarViewlet"
+        permission="zope2.View"
+        layer="ploneintranet.theme.interfaces.IThemeSpecific"
+      />
+
+
 </configure>
