+++ src/ploneintranet/layout/viewlets/personal_bar.pt
